+++ nix/package.nix
@@ -14,7 +14,7 @@
 }:
 
 let
-  version = "0.4.2-alpha";
+  version = "0.4.4-alpha";
 
   frontend = buildNpmPackage {
     pname = "hlidskjalf-frontend";